+++ projects/minecraft/v1.19.2/fabric/src/main/java/kittify/fabric/KittifyFabric.java
@@ -1,0 +1,11 @@
+package kittify.fabric;
+
+import kittify.Kittify;
+import net.fabricmc.api.ModInitializer;
+
+public class KittifyFabric implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        Kittify.LOGGER.info(Kittify.NAME + " is on Fabric.");
+    }
+}