+++ projects/minecraft/v1.19.2/forge/src/main/java/kittify/forge/KittifyForge.java
@@ -1,0 +1,11 @@
+package kittify.forge;
+
+import kittify.Kittify;
+import net.minecraftforge.fml.common.Mod;
+
+@Mod(Kittify.MOD_ID)
+public class KittifyForge {
+    public KittifyForge() {
+        Kittify.LOGGER.info(Kittify.NAME + " is on Forge.");
+    }
+}