+++ projects/minecraft/v1.19.2/quilt/src/main/java/kittify/quilt/KittifyQuilt.java
@@ -1,0 +1,12 @@
+package kittify.quilt;
+
+import kittify.Kittify;
+import org.quiltmc.loader.api.ModContainer;
+import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
+
+public class KittifyQuilt implements ModInitializer {
+    @Override
+    public void onInitialize(ModContainer mod) {
+        Kittify.LOGGER.info(Kittify.NAME + " is on Quilt");
+    }
+}