+++ projects/minecraft/v1.19.2/vanilla/src/main/java/kittify/Kittify.java
@@ -1,0 +1,10 @@
+package kittify;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class Kittify {
+    public static final String MOD_ID = "kittify";
+    public static final String NAME = "Kittify";
+    public static final Logger LOGGER = LogManager.getLogger();
+}